fix(stt): use float32 instead of the auto-selected int8
Correction d'une hypothèse fausse du commit précédent. Mesuré sur cette
machine (Apple Silicon, `large-v3-turbo`, 2,66 s de parole FR, CPU) :

    auto → int8   4609 ms   RTF 1.73
    float32       2361 ms   RTF 0.89

CTranslate2 n'a pas de noyau int8 accéléré sur ce CPU : la quantification
ajoute une déquantification par couche au lieu d'économiser du calcul.
`auto` choisissait pourtant int8 — c'était donc déjà le cas en production
et cela explique la moitié des 4944 ms observés.

La transcription est identique dans les trois configurations testées.

Deux tests verrouillent le réglage et sa traversée jusqu'au moteur, pour
qu'un retour silencieux à `auto` ne repasse pas inaperçu.

+++ config.py
@@ -85,9 +85,12 @@ DEFAULT_STT_MODEL = "large-v3-turbo"
 DEFAULT_STT_FALLBACK_MODEL = "large-v3"
 DEFAULT_STT_LANGUAGE = "fr"
 DEFAULT_STT_DEVICE = "auto"
-# int8 sur CPU Apple Silicon : ~2 à 3 fois plus rapide que float32 pour une
-# perte négligeable sur des commandes courtes. « auto » retombait sur float32.
-DEFAULT_STT_COMPUTE_TYPE = "int8"
+# Mesuré sur Apple Silicon (large-v3-turbo, 2,66 s de parole FR, CPU) :
+#   auto → int8   4609 ms   |   float32   2361 ms
+# CTranslate2 n'a pas de noyau int8 accéléré ici : la quantification ajoute une
+# déquantification par couche au lieu d'économiser du calcul. « auto » choisit
+# pourtant int8, ce qui doublait le temps de transcription. On fixe float32.
+DEFAULT_STT_COMPUTE_TYPE = "float32"
 DEFAULT_STT_BEAM_SIZE = 1
 DEFAULT_STT_VAD_FILTER = False
 DEFAULT_TTS_ENGINE = "kokoro"
